Change semantic of agent's preprocessor
Before preprocessors have to return numpy array, now we ask it to create
an object which can be passed directly into the model.
In the simplest case it has to be a Variable placed to the same GPU, so,
default preprocessor does just that. But it allows to implement more
complex preprocessing, like handling multimodal conversion of data.